+++ src/base/conditions/BaseConditionRule.php
@@ -102,6 +102,14 @@ abstract class BaseConditionRule extends Component implements ConditionRuleInter
     /**
      * @inheritdoc
      */
+    public function getGroupLabel(): ?string
+    {
+        return null;
+    }
+
+    /**
+     * @inheritdoc
+     */
     public function getConfig(): array
     {
         $config = [